+++ docs/recursion/README.md
@@ -5,12 +5,12 @@
     1.  [Fibonacci Number](i/FibonacciNumber.md)
     2.  [A to the Power of B](i/AToThePowerOfB.md)
 
-2.  [Recursion II](II)
-    1.  [N Queens](II/Medium/NQueens)
-    2.  [Spiral Order Traverse I](II/Medium/SpiralOrderTraverseI)
-    3.  [Spiral Order Traverse II](II/Medium/SpiralOrderTraverseII)
-    4.  [Lowest Common Ancestor I](II/Medium/LowestCommonAncestorI)
-    5.  [Lowest Common Ancestor II](II/Medium/LowestCommonAncestorII)
-    6.  [Maximum Path Sum Binary Tree III](II/Medium/MaximumPathSumBinaryTreeIII)
-    7.  [Maximum Path Sum Binary Tree II](II/Hard/MaximumPathSumBinaryTreeII)
-    8.  [Binary Tree Path Sum to Target III](II/Medium/BinaryTreePathSumToTargetIII)
+2.  [Recursion II](ii/README.md)
+    1.  [N Queens](ii/NQueens.md)
+    2.  [Spiral Order Traverse I](ii/SpiralOrderTraverseI.md)
+    3.  [Spiral Order Traverse II](ii/SpiralOrderTraverseII.md)
+    4.  [Lowest Common Ancestor I](ii/LowestCommonAncestorI.md)
+    5.  [Lowest Common Ancestor II](ii/LowestCommonAncestorII.md)
+    6.  [Maximum Path Sum Binary Tree III](../../coollime/src/main/java/coollime/recursion/maximumpathsumbinarytreeiii/Solution.java)
+    7.  [Maximum Path Sum Binary Tree II](ii/MaximumPathSumBinaryTreeII.md)
+    8.  [Binary Tree Path Sum to Target III](ii/BinaryTreePathSumToTargetIII.md)
