+++ docs/recursion/ii/lca0.png
