+++ docs/recursion/ii/lca1.png
