+++ docs/recursion/ii/nqueens.png
