--- src/C/Recursion/II/README.md
@@ -1,10 +1,0 @@
-## Recursion II
-
-1.  [N Queens](Medium/NQueens)
-2.  [Spiral Order Traverse I](Medium/SpiralOrderTraverseI)
-3.  [Spiral Order Traverse II](Medium/SpiralOrderTraverseII)
-4.  [Lowest Common Ancestor I](Medium/LowestCommonAncestorI)
-5.  [Lowest Common Ancestor II](Medium/LowestCommonAncestorII)
-6.  [Maximum Path Sum Binary Tree III](Medium/MaximumPathSumBinaryTreeIII)
-7.  [Maximum Path Sum Binary Tree II](Hard/MaximumPathSumBinaryTreeII)
-8.  [Binary Tree Path Sum to Target III](Medium/BinaryTreePathSumToTargetIII)
